requirements.txt
flask
twilio
pandas
scikit-learn
matplotlib
joblib
flask
google-generativeai
joblib
numpy